refactor(p2p): remove AttestationPool interface, rename KvAttestationPool to AttestationPool
- Merge KvAttestationPool class into attestation_pool.ts and rename to AttestationPool
- Move TryAddResult type definition to the class file
- Delete kv_attestation_pool.ts and kv_attestation_pool.test.ts files
- Add AttestationPoolApi type alias for typing mocks and test implementations
- Update all imports across the codebase to use the renamed class
- Update MemPools and P2PClientDeps types to use AttestationPoolApi for flexibility
- Rename test file from kv_attestation_pool.test.ts to attestation_pool.test.ts
