+++ yarn-project/p2p/src/client/p2p_client.test.ts
@@ -25,8 +25,7 @@ import { type MockProxy, mock } from 'jest-mock-extended';
 
 import type { P2PConfig } from '../config.js';
 import type { P2PService } from '../index.js';
-import type { AttestationPool } from '../mem_pools/attestation_pool/attestation_pool.js';
-import { createTestAttestationPool } from '../mem_pools/attestation_pool/kv_attestation_pool.js';
+import { type AttestationPool, createTestAttestationPool } from '../mem_pools/attestation_pool/attestation_pool.js';
 import type { MemPools } from '../mem_pools/interface.js';
 import { AztecKVTxPool } from '../mem_pools/tx_pool/aztec_kv_tx_pool.js';
 import type { TxPool } from '../mem_pools/tx_pool/index.js';
