+++ yarn-project/p2p/src/mem_pools/attestation_pool/attestation_pool.ts
@@ -1,5 +1,17 @@
-import type { SlotNumber } from '@aztec/foundation/branded-types';
-import type { BlockProposal, CheckpointAttestation, CheckpointProposalCore } from '@aztec/stdlib/p2p';
+import { IndexWithinCheckpoint, SlotNumber } from '@aztec/foundation/branded-types';
+import { Fr } from '@aztec/foundation/curves/bn254';
+import { toArray } from '@aztec/foundation/iterable';
+import { createLogger } from '@aztec/foundation/log';
+import type { AztecAsyncKVStore, AztecAsyncMap, AztecAsyncMultiMap } from '@aztec/kv-store';
+import {
+  BlockProposal,
+  CheckpointAttestation,
+  CheckpointProposal,
+  type CheckpointProposalCore,
+} from '@aztec/stdlib/p2p';
+import { type TelemetryClient, getTelemetryClient } from '@aztec/telemetry-client';
+
+import { PoolInstrumentation, PoolName, type PoolStatsCallback } from '../instrumentation.js';
 
 /** Result of trying to add an item (proposal or attestation) to the pool */
 export type TryAddResult = {
@@ -11,13 +23,131 @@ export type TryAddResult = {
   totalForPosition: number;
 };
 
+export const MAX_PROPOSALS_PER_SLOT = 5;
+export const MAX_PROPOSALS_PER_POSITION = 3;
+export const ATTESTATION_CAP_BUFFER = 10;
+
+/** Public API interface for attestation pools. Used for typing mocks and test implementations. */
+export type AttestationPoolApi = Pick<
+  AttestationPool,
+  | 'tryAddBlockProposal'
+  | 'getBlockProposal'
+  | 'tryAddCheckpointProposal'
+  | 'getCheckpointProposal'
+  | 'addCheckpointAttestations'
+  | 'tryAddCheckpointAttestation'
+  | 'deleteOlderThan'
+  | 'getCheckpointAttestationsForSlot'
+  | 'getCheckpointAttestationsForSlotAndProposal'
+  | 'isEmpty'
+>;
+
 /**
- * An Attestation Pool contains attestations collected by a validator
+ * Pool for storing attestations and proposals collected by a validator.
  *
- * Attestations that are observed via the p2p network are stored for requests
+ * Attestations and proposals observed via the p2p network are stored for requests
  * from the validator to produce a block, or to serve to other peers.
  */
-export interface AttestationPool {
+export class AttestationPool {
+  private metrics: PoolInstrumentation<CheckpointAttestation>;
+
+  // Checkpoint attestations from attestation key (slot-proposalId-signer) to serialized CheckpointAttestation
+  // Keys are lexicographically sortable allowing range queries by slot or by (slot, proposalId)
+  private checkpointAttestations: AztecAsyncMap<string, Buffer>;
+
+  // Checkpoint proposals from proposal archive to serialized CheckpointProposal
+  private checkpointProposals: AztecAsyncMap<string, Buffer>;
+
+  // Checkpoint proposals indexed by slot for querying all proposals in a slot
+  // Key: slot number, Value: proposal archive strings
+  private checkpointProposalsForSlot: AztecAsyncMultiMap<number, string>;
+
+  // Block proposals from proposal archive to serialized BlockProposal
+  private blockProposals: AztecAsyncMap<string, Buffer>;
+
+  // Block proposals indexed by slot and index-within-checkpoint for duplicate detection
+  // Key: (slot << 10) | indexWithinCheckpoint, Value: archive string
+  private blockProposalsForSlotAndIndex: AztecAsyncMultiMap<number, string>;
+
+  constructor(
+    private store: AztecAsyncKVStore,
+    telemetry: TelemetryClient = getTelemetryClient(),
+    private log = createLogger('aztec:attestation_pool'),
+  ) {
+    // Initialize block proposal storage
+    this.blockProposals = store.openMap('proposals');
+    this.blockProposalsForSlotAndIndex = store.openMultiMap('block_proposals_for_slot_and_index');
+
+    // Initialize checkpoint attestations storage
+    this.checkpointAttestations = store.openMap('checkpoint_attestations');
+
+    // Initialize checkpoint proposal storage
+    this.checkpointProposals = store.openMap('checkpoint_proposals');
+    this.checkpointProposalsForSlot = store.openMultiMap('checkpoint_proposals_for_slot');
+
+    this.metrics = new PoolInstrumentation(telemetry, PoolName.ATTESTATION_POOL, this.poolStats);
+  }
+
+  private poolStats: PoolStatsCallback = async () => {
+    return {
+      itemCount: await this.checkpointAttestations.sizeAsync(),
+    };
+  };
+
+  /** Returns whether the pool is empty. */
+  public async isEmpty(): Promise<boolean> {
+    for await (const _ of this.checkpointAttestations.entriesAsync()) {
+      return false;
+    }
+    for await (const _ of this.blockProposals.entriesAsync()) {
+      return false;
+    }
+    return true;
+  }
+
+  private getProposalKey(slot: number | bigint | Fr | string, proposalId: Fr | string | Buffer): string {
+    const slotStr = typeof slot === 'string' ? slot : new Fr(slot).toString();
+    const proposalIdStr =
+      typeof proposalId === 'string'
+        ? proposalId
+        : Buffer.isBuffer(proposalId)
+          ? Fr.fromBuffer(proposalId).toString()
+          : proposalId.toString();
+
+    return `${slotStr}-${proposalIdStr}`;
+  }
+
+  private getAttestationKey(slot: number | bigint | Fr | string, proposalId: Fr | string, address: string): string {
+    return `${this.getProposalKey(slot, proposalId)}-${address}`;
+  }
+
+  /** Returns range bounds for querying all attestations for a given slot. */
+  private getAttestationKeyRangeForSlot(slot: SlotNumber): { start: string; end: string } {
+    const slotStr = new Fr(slot).toString();
+    return { start: `${slotStr}-`, end: `${slotStr}-Z` }; // 'Z' sorts after any hex character
+  }
+
+  /** Returns range bounds for querying all attestations for a given (slot, proposalId). */
+  private getAttestationKeyRangeForProposal(slot: SlotNumber, proposalId: string): { start: string; end: string } {
+    const proposalKey = this.getProposalKey(slot, proposalId);
+    return { start: `${proposalKey}-`, end: `${proposalKey}-Z` };
+  }
+
+  /** Number of bits reserved for indexWithinCheckpoint in position keys. */
+  private static readonly INDEX_BITS = 10;
+  /** Maximum indexWithinCheckpoint value (2^10 - 1 = 1023). */
+  private static readonly MAX_INDEX = (1 << AttestationPool.INDEX_BITS) - 1;
+
+  /** Creates a position key for block proposals: (slot << 10) | indexWithinCheckpoint. */
+  private getBlockPositionKey(slot: number, indexWithinCheckpoint: number): number {
+    if (indexWithinCheckpoint > AttestationPool.MAX_INDEX) {
+      throw new Error(
+        `Value for indexWithinCheckpoint ${indexWithinCheckpoint} exceeds maximum ${AttestationPool.MAX_INDEX}`,
+      );
+    }
+    return (slot << AttestationPool.INDEX_BITS) | indexWithinCheckpoint;
+  }
+
   /**
    * Attempts to add a block proposal to the pool.
    *
@@ -29,7 +159,55 @@ export interface AttestationPool {
    * @param blockProposal - The block proposal to add
    * @returns Result indicating whether the proposal was added and duplicate detection info
    */
-  tryAddBlockProposal(blockProposal: BlockProposal): Promise<TryAddResult>;
+  public async tryAddBlockProposal(blockProposal: BlockProposal): Promise<TryAddResult> {
+    const proposalId = blockProposal.archive.toString();
+
+    // Check if already exists
+    const alreadyExists = await this.blockProposals.hasAsync(proposalId);
+    if (alreadyExists) {
+      const totalForPosition = await this.getBlockProposalCountForPosition(
+        blockProposal.slotNumber,
+        blockProposal.indexWithinCheckpoint,
+      );
+      return { added: false, alreadyExists: true, totalForPosition };
+    }
+
+    // Get current count for position and check cap, do not add if exceeded
+    const totalForPosition = await this.getBlockProposalCountForPosition(
+      blockProposal.slotNumber,
+      blockProposal.indexWithinCheckpoint,
+    );
+
+    if (totalForPosition >= MAX_PROPOSALS_PER_POSITION) {
+      return { added: false, alreadyExists: false, totalForPosition };
+    }
+
+    // Add the proposal
+    await this.addBlockProposal(blockProposal);
+
+    return { added: true, alreadyExists: false, totalForPosition: totalForPosition + 1 };
+  }
+
+  /** Gets the count of block proposals for a given position (slot, indexWithinCheckpoint). */
+  private getBlockProposalCountForPosition(
+    slot: SlotNumber,
+    indexWithinCheckpoint: IndexWithinCheckpoint,
+  ): Promise<number> {
+    const positionKey = this.getBlockPositionKey(slot, indexWithinCheckpoint);
+    return this.blockProposalsForSlotAndIndex.getValueCountAsync(positionKey);
+  }
+
+  private async addBlockProposal(blockProposal: BlockProposal): Promise<void> {
+    await this.store.transactionAsync(async () => {
+      const proposalId = blockProposal.archive.toString();
+      // Strip signedTxs before storing to avoid persisting full tx data
+      await this.blockProposals.set(proposalId, blockProposal.withoutSignedTxs().toBuffer());
+
+      // Index by slot and position for duplicate detection
+      const positionKey = this.getBlockPositionKey(blockProposal.slotNumber, blockProposal.indexWithinCheckpoint);
+      await this.blockProposalsForSlotAndIndex.set(positionKey, proposalId);
+    });
+  }
 
   /**
    * Get block proposal by its ID.
@@ -38,7 +216,18 @@ export interface AttestationPool {
    *
    * @return The block proposal if it exists, otherwise undefined.
    */
-  getBlockProposal(id: string): Promise<BlockProposal | undefined>;
+  public async getBlockProposal(id: string): Promise<BlockProposal | undefined> {
+    const buffer = await this.blockProposals.getAsync(id);
+    try {
+      if (buffer && buffer.length > 0) {
+        return BlockProposal.fromBuffer(buffer);
+      }
+    } catch {
+      return undefined;
+    }
+
+    return undefined;
+  }
 
   /**
    * Attempts to add a checkpoint proposal to the pool.
@@ -55,7 +244,37 @@ export interface AttestationPool {
    * @param proposal - The checkpoint proposal core to add
    * @returns Result indicating whether the proposal was added and duplicate detection info
    */
-  tryAddCheckpointProposal(proposal: CheckpointProposalCore): Promise<TryAddResult>;
+  public async tryAddCheckpointProposal(proposal: CheckpointProposalCore): Promise<TryAddResult> {
+    const proposalId = proposal.archive.toString();
+
+    // Check if already exists
+    const alreadyExists = await this.checkpointProposals.hasAsync(proposalId);
+    if (alreadyExists) {
+      const totalForPosition = await this.checkpointProposalsForSlot.getValueCountAsync(proposal.slotNumber);
+      return { added: false, alreadyExists: true, totalForPosition };
+    }
+
+    // Get current count for slot and check cap
+    const totalForPosition = await this.checkpointProposalsForSlot.getValueCountAsync(proposal.slotNumber);
+    if (totalForPosition >= MAX_PROPOSALS_PER_SLOT) {
+      return { added: false, alreadyExists: false, totalForPosition };
+    }
+
+    // Add the proposal if cap not exceeded
+    await this.addCheckpointProposal(proposal);
+
+    return { added: true, alreadyExists: false, totalForPosition: totalForPosition + 1 };
+  }
+
+  private async addCheckpointProposal(proposal: CheckpointProposalCore): Promise<void> {
+    await this.store.transactionAsync(async () => {
+      const slotKey = proposal.slotNumber;
+      const proposalId = proposal.archive.toString();
+
+      await this.checkpointProposalsForSlot.set(slotKey, proposalId);
+      await this.checkpointProposals.set(proposalId, proposal.toBuffer());
+    });
+  }
 
   /**
    * Get checkpoint proposal by its ID.
@@ -66,14 +285,141 @@ export interface AttestationPool {
    * @param id - The ID of the checkpoint proposal to retrieve (proposal.archive)
    * @return The checkpoint proposal core if it exists, otherwise undefined.
    */
-  getCheckpointProposal(id: string): Promise<CheckpointProposalCore | undefined>;
+  public async getCheckpointProposal(id: string): Promise<CheckpointProposalCore | undefined> {
+    const buffer = await this.checkpointProposals.getAsync(id);
+    try {
+      if (buffer && buffer.length > 0) {
+        return CheckpointProposal.fromBuffer(buffer);
+      }
+    } catch {
+      return undefined;
+    }
+
+    return undefined;
+  }
 
   /**
-   * Add checkpoint attestations to the pool
+   * Add checkpoint attestations to the pool.
    *
    * @param attestations - Checkpoint attestations to add into the pool
    */
-  addCheckpointAttestations(attestations: CheckpointAttestation[]): Promise<void>;
+  public async addCheckpointAttestations(attestations: CheckpointAttestation[]): Promise<void> {
+    await this.store.transactionAsync(async () => {
+      for (const attestation of attestations) {
+        const slotNumber = attestation.payload.header.slotNumber;
+        const proposalId = attestation.archive;
+        const sender = attestation.getSender();
+
+        // Skip attestations with invalid signatures
+        if (!sender) {
+          this.log.warn(`Skipping checkpoint attestation with invalid signature for slot ${slotNumber}`, {
+            signature: attestation.signature.toString(),
+            slotNumber,
+            proposalId,
+          });
+          continue;
+        }
+
+        const address = sender.toString();
+
+        await this.checkpointAttestations.set(
+          this.getAttestationKey(slotNumber, proposalId, address),
+          attestation.toBuffer(),
+        );
+
+        this.log.verbose(`Added checkpoint attestation for slot ${slotNumber} from ${address}`, {
+          signature: attestation.signature.toString(),
+          slotNumber,
+          address,
+          proposalId,
+        });
+      }
+    });
+  }
+
+  /**
+   * Get all checkpoint attestations for a given slot.
+   *
+   * @param slot - The slot to query
+   * @return CheckpointAttestations
+   */
+  public async getCheckpointAttestationsForSlot(slot: SlotNumber): Promise<CheckpointAttestation[]> {
+    const range = this.getAttestationKeyRangeForSlot(slot);
+    const attestations: CheckpointAttestation[] = [];
+
+    for await (const [_, buf] of this.checkpointAttestations.entriesAsync(range)) {
+      attestations.push(CheckpointAttestation.fromBuffer(buf));
+    }
+
+    return attestations;
+  }
+
+  /**
+   * Get checkpoint attestations for slot and given proposal.
+   *
+   * @param slot - The slot to query
+   * @param proposalId - The proposal to query
+   * @return CheckpointAttestations
+   */
+  public async getCheckpointAttestationsForSlotAndProposal(
+    slot: SlotNumber,
+    proposalId: string,
+  ): Promise<CheckpointAttestation[]> {
+    const range = this.getAttestationKeyRangeForProposal(slot, proposalId);
+    const attestations: CheckpointAttestation[] = [];
+
+    for await (const [_, buf] of this.checkpointAttestations.entriesAsync(range)) {
+      attestations.push(CheckpointAttestation.fromBuffer(buf));
+    }
+
+    return attestations;
+  }
+
+  /**
+   * Delete all pool data (attestations, proposals) older than the given slot.
+   *
+   * @param oldestSlot - The oldest slot to keep.
+   */
+  public async deleteOlderThan(oldestSlot: SlotNumber): Promise<void> {
+    let numberOfAttestations = 0;
+    let numberOfCheckpointProposals = 0;
+    let numberOfBlockPositions = 0;
+
+    await this.store.transactionAsync(async () => {
+      // Delete checkpoint attestations with slot < oldestSlot
+      // Attestation keys start with Fr(slot).toString(), so we use end bound of Fr(oldestSlot).toString()
+      const attestationEndKey = new Fr(oldestSlot).toString();
+      for await (const key of this.checkpointAttestations.keysAsync({ end: attestationEndKey })) {
+        await this.checkpointAttestations.delete(key);
+        numberOfAttestations++;
+      }
+
+      // Delete checkpoint proposals for slots < oldestSlot, using checkpointProposalsForSlot as index
+      for await (const slot of this.checkpointProposalsForSlot.keysAsync({ end: oldestSlot })) {
+        const proposalIds = await toArray(this.checkpointProposalsForSlot.getValuesAsync(slot));
+        for (const proposalId of proposalIds) {
+          await this.checkpointProposals.delete(proposalId);
+          numberOfCheckpointProposals++;
+        }
+        await this.checkpointProposalsForSlot.delete(slot);
+      }
+
+      // Delete block proposal position index for slots < oldestSlot
+      // Key format: (slot << INDEX_BITS) | indexWithinCheckpoint
+      const blockPositionEndKey = oldestSlot << AttestationPool.INDEX_BITS;
+      for await (const positionKey of this.blockProposalsForSlotAndIndex.keysAsync({ end: blockPositionEndKey })) {
+        await this.blockProposalsForSlotAndIndex.delete(positionKey);
+        numberOfBlockPositions++;
+      }
+    });
+
+    this.log.verbose(`Deleted old pool data`, {
+      oldestSlot,
+      numberOfAttestations,
+      numberOfCheckpointProposals,
+      numberOfBlockPositions,
+    });
+  }
 
   /**
    * Attempts to add a checkpoint attestation to the pool.
@@ -87,32 +433,59 @@ export interface AttestationPool {
    * @param committeeSize - Committee size for the attestation's slot
    * @returns Result indicating whether the attestation was added and existence info
    */
-  tryAddCheckpointAttestation(attestation: CheckpointAttestation, committeeSize: number): Promise<TryAddResult>;
+  public async tryAddCheckpointAttestation(
+    attestation: CheckpointAttestation,
+    committeeSize: number,
+  ): Promise<TryAddResult> {
+    const slotNumber = attestation.payload.header.slotNumber;
+    const proposalId = attestation.archive.toString();
+    const sender = attestation.getSender();
 
-  /**
-   * Delete all pool data (attestations, proposals) older than the given slot
-   *
-   * @param slot - The oldest slot to keep.
-   */
-  deleteOlderThan(slot: SlotNumber): Promise<void>;
+    if (!sender) {
+      return { added: false, alreadyExists: false, totalForPosition: 0 };
+    }
 
-  /**
-   * Get all checkpoint attestations for a given slot
-   *
-   * @param slot - The slot to query
-   * @return CheckpointAttestations
-   */
-  getCheckpointAttestationsForSlot(slot: SlotNumber): Promise<CheckpointAttestation[]>;
+    const key = this.getAttestationKey(slotNumber, proposalId, sender.toString());
+    const alreadyExists = await this.checkpointAttestations.hasAsync(key);
 
-  /**
-   * Get checkpoint attestations for slot and given proposal
-   *
-   * @param slot - The slot to query
-   * @param proposalId - The proposal to query
-   * @return CheckpointAttestations
-   */
-  getCheckpointAttestationsForSlotAndProposal(slot: SlotNumber, proposalId: string): Promise<CheckpointAttestation[]>;
+    if (alreadyExists) {
+      const total = await this.getAttestationCount(slotNumber, proposalId);
+      return { added: false, alreadyExists: true, totalForPosition: total };
+    }
 
-  /** Returns whether the pool is empty. */
-  isEmpty(): Promise<boolean>;
+    const limit = committeeSize + ATTESTATION_CAP_BUFFER;
+    const currentCount = await this.getAttestationCount(slotNumber, proposalId);
+
+    if (currentCount >= limit) {
+      return { added: false, alreadyExists: false, totalForPosition: currentCount };
+    }
+
+    await this.checkpointAttestations.set(key, attestation.toBuffer());
+
+    this.log.verbose(`Added checkpoint attestation for slot ${slotNumber} from ${sender.toString()}`, {
+      signature: attestation.signature.toString(),
+      slotNumber,
+      address: sender.toString(),
+      proposalId,
+    });
+
+    return { added: true, alreadyExists: false, totalForPosition: currentCount + 1 };
+  }
+
+  /** Gets the count of attestations for a given (slot, proposalId). */
+  private async getAttestationCount(slot: SlotNumber, proposalId: string): Promise<number> {
+    const range = this.getAttestationKeyRangeForProposal(slot, proposalId);
+    let count = 0;
+    for await (const _ of this.checkpointAttestations.keysAsync(range)) {
+      count++;
+    }
+    return count;
+  }
+}
+
+/** Creates an AttestationPool backed by a temporary store for testing. */
+export async function createTestAttestationPool(telemetry?: TelemetryClient): Promise<AttestationPool> {
+  const { openTmpStore } = await import('@aztec/kv-store/lmdb-v2');
+  const store = await openTmpStore('test-attestation-pool');
+  return new AttestationPool(store, telemetry);
 }
