+++ yarn-project/p2p/src/mem_pools/attestation_pool/index.ts
@@ -1,2 +1,9 @@
-export * from './attestation_pool.js';
-export * from './kv_attestation_pool.js';
+export {
+  AttestationPool,
+  type AttestationPoolApi,
+  type TryAddResult,
+  createTestAttestationPool,
+  MAX_PROPOSALS_PER_SLOT,
+  MAX_PROPOSALS_PER_POSITION,
+  ATTESTATION_CAP_BUFFER,
+} from './attestation_pool.js';
