+++ yarn-project/p2p/src/mem_pools/index.ts
@@ -1,3 +1,3 @@
-export { type AttestationPool } from './attestation_pool/attestation_pool.js';
+export { AttestationPool, type AttestationPoolApi } from './attestation_pool/attestation_pool.js';
 export { type MemPools } from './interface.js';
 export { type TxPool } from './tx_pool/tx_pool.js';
