+++ yarn-project/p2p/src/services/libp2p/libp2p_service.test.ts
@@ -26,12 +26,11 @@ import { TopicValidatorResult } from '@libp2p/interface';
 import { type MockProxy, mock } from 'jest-mock-extended';
 
 import { type P2PConfig, p2pConfigMappings } from '../../config.js';
-import type { AttestationPool } from '../../mem_pools/attestation_pool/attestation_pool.js';
 import {
-  KvAttestationPool,
+  AttestationPool,
   MAX_PROPOSALS_PER_POSITION,
   MAX_PROPOSALS_PER_SLOT,
-} from '../../mem_pools/attestation_pool/kv_attestation_pool.js';
+} from '../../mem_pools/attestation_pool/attestation_pool.js';
 import type { MemPools } from '../../mem_pools/interface.js';
 import type { TxPool } from '../../mem_pools/tx_pool/tx_pool.js';
 import type { PubSubLibp2p } from '../../util.js';
@@ -376,7 +375,7 @@ describe('LibP2PService', () => {
   });
 
   describe('processBlockFromPeer', () => {
-    let attestationPool: KvAttestationPool;
+    let attestationPool: AttestationPool;
     let mockTxPool: MockProxy<TxPool>;
     let mockEpochCache: MockProxy<EpochCacheInterface>;
     let signer: Secp256k1Signer;
@@ -388,7 +387,7 @@ describe('LibP2PService', () => {
 
     beforeEach(() => {
       signer = Secp256k1Signer.random();
-      attestationPool = new KvAttestationPool(openTmpStore(true));
+      attestationPool = new AttestationPool(openTmpStore(true));
       mockTxPool = mock<TxPool>();
       mockTxPool.markTxsAsNonEvictable.mockResolvedValue();
 
@@ -614,7 +613,7 @@ describe('LibP2PService', () => {
   });
 
   describe('handleGossipedCheckpointProposal', () => {
-    let attestationPool: KvAttestationPool;
+    let attestationPool: AttestationPool;
     let mockTxPool: MockProxy<TxPool>;
     let mockEpochCache: MockProxy<EpochCacheInterface>;
     let signer: Secp256k1Signer;
@@ -627,7 +626,7 @@ describe('LibP2PService', () => {
 
     beforeEach(() => {
       signer = Secp256k1Signer.random();
-      attestationPool = new KvAttestationPool(openTmpStore(true));
+      attestationPool = new AttestationPool(openTmpStore(true));
       mockTxPool = mock<TxPool>();
       mockTxPool.markTxsAsNonEvictable.mockResolvedValue();
 
@@ -999,7 +998,7 @@ function createTestLibP2PService(options: CreateTestLibP2PServiceOptions): TestL
     peerManager,
     node,
     archiver = mock<L2BlockSource & ContractDataSource>(),
-    attestationPool = new KvAttestationPool(openTmpStore(true)),
+    attestationPool = new AttestationPool(openTmpStore(true)),
     txPool = mock<TxPool>(),
     epochCache = mock<EpochCacheInterface>(),
   } = options;
@@ -1018,7 +1017,7 @@ function createTestLibP2PService(options: CreateTestLibP2PServiceOptions): TestL
 function createTestLibP2PServiceWithPools(
   mockPeerManager: MockProxy<PeerManagerInterface>,
   reportMessageValidationResultSpy: jest.Mock,
-  attestationPool: KvAttestationPool,
+  attestationPool: AttestationPool,
   mockTxPool: MockProxy<TxPool>,
   mockEpochCache: MockProxy<EpochCacheInterface>,
 ): TestLibP2PService {
