+++ yarn-project/p2p/src/services/reqresp/protocols/block_txs/block_txs_handler.ts
@@ -4,7 +4,8 @@ import { TxArray } from '@aztec/stdlib/tx';
 
 import type { PeerId } from '@libp2p/interface';
 
-import type { AttestationPool, TxPool } from '../../../../mem_pools/index.js';
+import type { AttestationPoolApi } from '../../../../mem_pools/attestation_pool/attestation_pool.js';
+import type { TxPool } from '../../../../mem_pools/index.js';
 import type { ReqRespSubProtocolHandler } from '../../interface.js';
 import { ReqRespStatus, ReqRespStatusError } from '../../status.js';
 import { BitVector } from './bitvector.js';
@@ -18,7 +19,7 @@ import { BlockTxsRequest, BlockTxsResponse } from './block_txs_reqresp.js';
  * @returns the BlockTxs request handler
  */
 export function reqRespBlockTxsHandler(
-  attestationPool: AttestationPool,
+  attestationPool: AttestationPoolApi,
   archiver: L2BlockSource,
   txPool: TxPool,
 ): ReqRespSubProtocolHandler {
